Move user language into UserSettings
